Apollo

Documentation for Apollo.

Utility Methods

Apollo.catlayers — Method

catlayers(x::AbstractRasterStack, dim)

Concatenate the layers of x along the dimension given by dim.

Apollo.dropobs — Method

dropobs(x::AbstractArray)

Remove the observation dimension from the tensor x.

Apollo.folddims — Method

folddims(f, xs::AbstractRaster; dims=Band)

Apply the reducer function f to all non-missing elements in each slice of x WRT dims.

Arguments

f: A function that reduces an array of values to a singular value.

x: An AbstractRaster over which we want to fold.

dims: The dimension used to generate each slice that is passed to f.

Example

julia> μ = folddims(mean, raster, dims=Band)
6-element Vector{Float32}:
 0.09044644
 0.23737456
 0.30892986
 0.33931717
 0.16186203
 0.076255515

Apollo.foldlayers — Method

foldlayers(f, x::AbstractRasterStack)

Apply the reducer f to all non-missing elements in each layer of x.

Apollo.ones_like — Method

ones_like(x::AbstractArray)

Construct an array of ones with the same size and element type as x.

Apollo.putdim — Method

putdim(raster::AbstractRaster, dims)

Add the provided singleton dim(s) to raster. Does nothing if dims is already present.

Example

julia> r = Raster(rand(512,512), (X,Y));

julia> putdim(r, Band)
╭─────────────────────────────╮
│ 512×512×1 Raster{Float64,3} │
├─────────────────────────────┴─────────────────────────────── dims ┐
  ↓ X   ,
  → Y   ,
  ↗ Band Sampled{Int64} Base.OneTo(1) ForwardOrdered Regular Points
├─────────────────────────────────────────────────────────── raster ┤
  extent: Extent(X = (1, 512), Y = (1, 512), Band = (1, 1))

└───────────────────────────────────────────────────────────────────┘
[:, :, 1]
 0.107662  0.263251    0.786834  0.334663  …  0.316804   0.709557    0.478199
 0.379863  0.532268    0.635206  0.33514      0.402433   0.413602    0.657538
 0.129775  0.283808    0.327946  0.727027     0.685844   0.847777    0.435326
 0.73348   0.00705636  0.178885  0.381932     0.146575   0.310242    0.159852
 ⋮                                         ⋱                         ⋮
 0.330857  0.52704     0.888379  0.811084  …  0.0660687  0.00230472  0.448761
 0.698654  0.510846    0.916446  0.621061     0.23648    0.510697    0.113338
 0.600629  0.116626    0.567983  0.174267     0.089853   0.443758    0.667935

Apollo.putobs — Method

putobs(x::AbstractArray)

Add an N+1 obervation dimension of size 1 to the tensor x.

Apollo.stackobs — Method

stackobs(x...)
stackobs(x::AbstractVector)

Stack the elements in x as if they were observations in a batch. If x is an AbstractArray, elements will be concatenated along the Nth dimension. Other data types will simply be placed in a Vector in the same order as they were received. Special attention is paid to a Vector of Tuples, where each tuple represents a single observation, such as a feature/label pair. In this case, the tuples will first be unzipped, before their constituent elements are then stacked as usual.

Example

julia> stackobs(1, 2, 3, 4, 5)
5-element Vector{Int64}:
 1
 2
 3
 4
 5

julia> stackobs([(1, :a), (2, :b), (3, :c)])
([1, 2, 3], [:a, :b, :c])

julia> stackobs([rand(256, 256, 3, 1) for _ in 1:10]...) |> size
(256, 256, 3, 10)

julia> xs = [rand(256, 256, 3, 1) for _ in 1:10];

julia> ys = [rand(256, 256, 1, 1) for _ in 1:10];

julia> data = collect(zip(xs, ys));

julia> stackobs(data) .|> size
((256, 256, 3, 10), (256, 256, 1, 10))

Apollo.todevice — Method

todevice(x)

Copy x to the GPU.

Apollo.unzip — Method

unzip(x::AbstractVector{<:Tuple})

The reverse of zip.

Example

julia> zip([1, 2, 3], [:a, :b, :c]) |> collect |> unzip
([1, 2, 3], [:a, :b, :c])

Apollo.vec2array — Method

vec2array(x::AbstractVector, to::AbstractArray, dim::Int)

Reshape the vector x to have the same number of dimensions as to. Missing dimensions are added as singletons while the dimension corresponding to dim will be filled with the values of x.

Apollo.zeros_like — Method

zeros_like(x::AbstractArray)

Construct an array of zeros with the same size and element type as x.

Apollo.mosaicview — Method

mosaicview(size::Tuple{Int,Int}, imgs...)

Plot a mosaic of images with size (rows, cols).

Data

Apollo.DType — Type

The super-type of all data types.

Apollo.Image — Type

Image([name])

Represents a type consisting of image data.

Apollo.Mask — Type

Mask([name])

Represents an instance of mask data.

Apollo.ObsView — Type

ObsView(data, indices)

Construct an iterator over data

Apollo.dropobs — Method

dropobs(data, obs::AbstractVector{Int})

Remove all observations from data whose index corresponds to those in obs.

Apollo.filterobs — Method

filterobs(f, data)

Remove all observations from data for which f returns false.

Apollo.mapobs — Method

mapobs(f, data)

Lazily apply the function f to each element in data.

Apollo.repeatobs — Method

repeatobs(data, n::Int)

Create a new view which iterates over every element in data n times.

Apollo.sampleobs — Method

sampleobs([rng], data, n)

Randomly sample n elements from data without replacement. rng may be optionally provided for reproducible results.

Apollo.shuffleobs — Method

shuffleobs([rng], data)

Randomly shuffle the elements of data. Provide rng for reproducible results.

Apollo.splitobs — Method

splitobs(data; at=0.8, shuffle=true)

Return a set of indices that splits the given observations according to the given break points.

Arguments

data: Any type that implements Base.length().

at: The fractions at which to split data.

shuffle: If true, shuffles the indices before splitting.

Example

julia> splitobs(1:100, at=(0.7, 0.2), shuffle=false)
3-element Vector{Vector{Int64}}:
 [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  61, 62, 63, 64, 65, 66, 67, 68, 69, 70]
 [71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90]
 [91, 92, 93, 94, 95, 96, 97, 98, 99, 100]

Apollo.takeobs — Method

takeobs(data, obs::AbstractVector{Int})

Take all observations from data whose index corresponds to obs while removing everything else.

Apollo.zipobs — Method

zipobs(data...)

Create a new iterator where the elements of each iterator in data are returned as a tuple.

Example

julia> zipobs(1:5, 41:45, rand([:cat, :dog], 5)) |> collect
5-element Vector{Any}:
 (1, 41, :dog)
 (2, 42, :cat)
 (3, 43, :cat)
 (4, 44, :dog)
 (5, 45, :cat)

Apollo.TileSampler — Type

TileSampler(raster, tilesize::Int; stride=tilesize)

An object that iterates over tiles cut from a given raster.

Parameters

raster: An AbstractRaster or AbstractRasterStack to be cut into tiles.

tilesize: The size (width and height) of the tiles.

stride: The horizontal and vertical distance between adjacent tiles.

Apollo.crop — Function

crop(x::AbstractArray, size, ul=(1,1))

Crop a tile equal to size out of x with an upper-left corner defined by ul.

Apollo.raster — Method

raster(tensor::AbstractArray, dims::Tuple; missingval=0)

Restore the raster dimensions given by dims to the provided tensor. The final dimension of tensor, which is assumed to be the observation dimension, will be dropped.

Apollo.tensor — Method

tensor([precision], xs...; kwargs...)
tensor([precision], x::AbstractArray; kwargs...)
tensor([precision], x::AbstractRasterStack; layerdim=Band)

Convert one or more arrays to a tensor with an element type of precision. AbstractRasters will be reshaped as necessary to enforce a dimension order of (X,Y,Z,Band,Ti) before adding an observation dimension. Multiple arrays will be concatenated along the observation dimension after being converted to tensors.

Parameters

precision: Any AbstractFloat to use as the tensor's type (default = Float32).

x: One or more AbstractArrays to be turned into tensors.

layerdim: AbstractRasterStacks will have their layers concatenated along this dimension

before being turned into tensors.

Apollo.transform — Method

transform(t::AbstractTransform, dtype::DType, x)
transform(t::AbstractTransform, dtypes::Tuple, x::Tuple)

Apply the transformation t to the input x with data type dtype.

Metrics

Apollo.AbstractMetric — Type

Abstract supertype of all evaluation metrics.

Apollo.Metric — Type

Metric(measure::AbstractMetric)

Construct a Metric object to track the state for the given AbstractMetric.

Apollo.compute — Function

compute(m::AbstractMetric, state)
compute(m::Metric)

Compute the performance measure from the current state.

Apollo.evaluate — Method

evaluate(model, data, measures...)

Evaluate the model's performance on the provided data.

Parameters

model: A callable that takes a single batch from data and returns a tuple of the form (ŷ, y).

data: An iterable of (x, y) values.

measures: A set of AbstractMetrics to use for evaluating model.

Returns

A NamedTuple containing the performance metrics for the given model.

Example

evaluate(DataLoader((xsampler, ysampler)), Accuracy(), MIoU(2)) do (x, y)
    ŷ = model(x) |> Flux.sigmoid
    return (ŷ, y)
end

Apollo.init — Function

init(m::AbstractMetric)

Returns the initial state of the performance measure, which will be subsequently updated for each mini-batch of labels and predictions.

Apollo.name — Method

name(m::AbstractMetric)
name(m::Metric)

Human readable name of the given performance measure.

Apollo.reset! — Method

reset!(metric::Metric)

Reset the metric's state.

Apollo.update — Function

update(m::AbstractMetric, state, ŷ, y)

Return the new state for the given batch of labels and predictions.

Apollo.update! — Method

update!(metric::Metric, ŷ, y)
update!(tracker::Tracker, ŷ, y)

Update the metric state for the next batch of labels and predictions.

Apollo.epoch! — Method

epoch!(tracker::Tracker)

Store the metric value for the current epoch and reset the state.

Apollo.step! — Method

step!(tracker::Tracker, ŷ, y)

Update the metric's state for the current batch of predictions and labels.

Models

Apollo.UNet — Method

UNet(;encoder=StandardEncoder(batch_norm=true), input=nothing, channels=3, nclasses=1, batch_norm=true)

Construct a UNet model.

Keywords

encoder: The encoder to use for the UNet model. Defaults to the standard encoder.

input: The input block, which defaults to two convolutional layers as with standard UNet.

channels: The number of input channels to use when input=nothing. Ignored when input is specified.

nclasses: The number of output channels produced by the head.

batch_norm: Use batch normalization after each convolutional layer (default=true).