Apollo

Documentation for Apollo.

Apollo.AbstractMetric

Apollo.Metric

Apollo.ObsView

Apollo.UNet

Apollo.compute

Apollo.crop

Apollo.dropobs

Apollo.epoch!

Apollo.evaluate

Apollo.filterobs

Apollo.folddims

Apollo.init

Apollo.mapobs

Apollo.mosaicview

Apollo.name

Apollo.putdim

Apollo.repeatobs

Apollo.reset!

Apollo.sampleobs

Apollo.shuffleobs

Apollo.splitobs

Apollo.step!

Apollo.takeobs

Apollo.update

Apollo.update!

Apollo.vec2array

Apollo.zipobs

Apollo.AbstractMetric — Type

Abstract supertype of all evaluation metrics.

Apollo.Metric — Type

Metric(measure::AbstractMetric)

Construct a Metric object to track the state for the given AbstractMetric.

Apollo.ObsView — Type

ObsView(data, indices)

Construct an iterator over data

Apollo.UNet — Method

UNet(;encoder=StandardEncoder(batch_norm=true), input=nothing, channels=3, nclasses=1, batch_norm=true)

Construct a UNet model.

Keywords

encoder: The encoder to use for the UNet model. Defaults to the standard encoder.

input: The input block, which defaults to two convolutional layers as with standard UNet.

channels: The number of input channels to use when input=nothing. Ignored when input is specified.

nclasses: The number of output channels produced by the head.

batch_norm: Use batch normalization after each convolutional layer (default=true).

Apollo.compute — Function

compute(m::AbstractMetric, state)
compute(m::Metric)

Compute the performance measure from the current state.

Apollo.crop — Function

crop(x::AbstractArray, size, ul=(1,1))

Crop a tile equal to size out of x with an upper-left corner defined by ul.

Apollo.dropobs — Method

dropobs(data, obs::AbstractVector{Int})

Remove all observations from data whose index corresponds to those in obs.

Apollo.epoch! — Method

epoch!(tracker::Tracker)

Store the metric value for the current epoch and reset the state.

Apollo.evaluate — Method

evaluate(model, data, measures...)

Evaluate the model's performance on the provided data.

Parameters

model: A callable that takes a single batch from data and returns a tuple of the form (ŷ, y).

data: An iterable of (x, y) values.

measures: A set of AbstractMetrics to use for evaluating model.

Returns

A NamedTuple containing the performance metrics for the given model.

Example

evaluate(DataLoader((xsampler, ysampler)), Accuracy(), MIoU(2)) do (x, y)
    ŷ = model(x) |> Flux.sigmoid
    return (ŷ, y)
end

Apollo.filterobs — Method

filterobs(f, data)

Remove all observations from data for which f returns false.

Apollo.folddims — Method

folddims(f, xs::AbstractRaster; dim=Band)

Reduce the collection of non-missing values to a singular value in each slice of x WRT dims.

Arguments

f: A function that reduces an array of values to a singular value.

x: An AbstractRaster over which we want to fold.

dims: The dimension used to generate each slice that is passed to f.

Example

julia> μ = folddims(mean, raster, dims=Band)
6-element Vector{Float32}:
 0.09044644
 0.23737456
 0.30892986
 0.33931717
 0.16186203
 0.076255515

Apollo.init — Function

init(m::AbstractMetric)

Returns the initial state of the performance measure, which will be subsequently updated for each mini-batch of labels and predictions.

Apollo.mapobs — Method

mapobs(f, data)

Lazily apply the function f to each element in data.

Apollo.mosaicview — Method

mosaicview(size::Tuple{Int,Int}, imgs...)

Plot a mosaic of images with size (rows, cols).

Apollo.name — Method

name(m::AbstractMetric)
name(m::Metric)

Human readable name of the given performance measure.

Apollo.putdim — Method

putdim(raster::AbstractRaster, dims::Tuple)
putdim(x::AbstractRaster, ::Type{Rasters.DD.Dimension})

Add the provided singleton dim(s) to the given raster. Does nothing if dim is already present.

Apollo.repeatobs — Method

repeatobs(data, n::Int)

Create a new view which iterates over every element in data n times.

Apollo.reset! — Method

reset!(metric::Metric)

Reset the metric's state.

Apollo.sampleobs — Method

sampleobs([rng], data, n)

Randomly sample n elements from data without replacement. rng may be optionally provided for reproducible results.

Apollo.shuffleobs — Method

shuffleobs([rng], data)

Randomly shuffle the elements of data. Provide rng for reproducible results.

Apollo.splitobs — Method

splitobs(data; at=0.8, shuffle=true)

Return a set of indices that splits the given observations according to the given break points.

Arguments

data: Any type that implements Base.length().

at: The fractions at which to split data.

shuffle: If true, shuffles the indices before splitting.

Example

julia> splitobs(1:100, at=(0.7, 0.2), shuffle=false)
3-element Vector{Vector{Int64}}:
 [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  61, 62, 63, 64, 65, 66, 67, 68, 69, 70]
 [71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90]
 [91, 92, 93, 94, 95, 96, 97, 98, 99, 100]

Apollo.step! — Method

step!(tracker::Tracker, ŷ, y)

Update the metric's state for the current batch of predictions and labels.

Apollo.takeobs — Method

takeobs(data, obs::AbstractVector{Int})

Take all observations from data whose index corresponds to obs while removing everything else.

Apollo.update — Function

update(m::AbstractMetric, state, ŷ, y)

Return the new state for the given batch of labels and predictions.

Apollo.update! — Method

update!(metric::Metric, ŷ, y)
update!(tracker::Tracker, ŷ, y)

Update the metric state for the next batch of labels and predictions.

Apollo.vec2array — Method

vec2array(x::AbstractVector, to::AbstractArray, dim::Int)

Reshape the vector x to have the same number of dimensions as to. Missing dimensions are added as singletons while the dimension corresponding to dim will be filled with the values of x.

Apollo.zipobs — Method

zipobs(data...)

Create a new iterator where the elements of each iterator in data are returned as a tuple.

Example

julia> zipobs(1:5, 41:45, rand([:cat, :dog], 5)) |> collect
5-element Vector{Any}:
 (1, 41, :dog)
 (2, 42, :cat)
 (3, 43, :cat)
 (4, 44, :dog)
 (5, 45, :cat)