Apollo

Documentation for Apollo.

Utility Methods

Apollo.binmask — Function

binmask(x)

Visualze x as a black/white binary mask.

Apollo.catlayers — Function

catlayers(x::AbstractRasterStack, dim)

Concatenate the layers of x along the dimension given by dim.

Apollo.folddims — Function

folddims(f, xs::AbstractRaster; dims=Band)

Apply the reducer function f to all non-missing elements in each slice of x WRT dims.

Arguments

f: A function that reduces an array of values to a singular value.

x: An AbstractRaster over which we want to fold.

dims: The dimension used to generate each slice that is passed to f.

Example

julia> μ = folddims(mean, raster, dims=Band)
6-element Vector{Float32}:
 0.09044644
 0.23737456
 0.30892986
 0.33931717
 0.16186203
 0.076255515

Apollo.foldlayers — Function

foldlayers(f, x::AbstractRasterStack)

Apply the reducer f to all non-missing elements in each layer of x.

Apollo.linear_stretch — Function

linear_stretch(x::AbstractArray{<:Real,3}, lb::Vector{<:Real}, ub::Vector{<:Real})

Perform a linear histogram stretch on x such that lb is mapped to 0 and ub is mapped to 1. Values outside the interval [lb, ub] will be clamped.

Apollo.mosaicview — Function

mosaicview(size::Tuple{Int,Int}, imgs...)

Plot a mosaic of images with size (rows, cols).

Apollo.ones_like — Function

ones_like(x::AbstractArray)

Construct an array of ones with the same size and element type as x.

Apollo.putdim — Function

putdim(raster::AbstractRaster, dims)

Add the provided singleton dim(s) to raster. Does nothing if dims is already present.

Example

julia> r = Raster(rand(512,512), (X,Y));

julia> putdim(r, Band)
╭─────────────────────────────╮
│ 512×512×1 Raster{Float64,3} │
├─────────────────────────────┴─────────────────────────────── dims ┐
  ↓ X   ,
  → Y   ,
  ↗ Band Sampled{Int64} Base.OneTo(1) ForwardOrdered Regular Points
├─────────────────────────────────────────────────────────── raster ┤
  extent: Extent(X = (1, 512), Y = (1, 512), Band = (1, 1))

└───────────────────────────────────────────────────────────────────┘
[:, :, 1]
 0.107662  0.263251    0.786834  0.334663  …  0.316804   0.709557    0.478199
 0.379863  0.532268    0.635206  0.33514      0.402433   0.413602    0.657538
 0.129775  0.283808    0.327946  0.727027     0.685844   0.847777    0.435326
 0.73348   0.00705636  0.178885  0.381932     0.146575   0.310242    0.159852
 ⋮                                         ⋱                         ⋮
 0.330857  0.52704     0.888379  0.811084  …  0.0660687  0.00230472  0.448761
 0.698654  0.510846    0.916446  0.621061     0.23648    0.510697    0.113338
 0.600629  0.116626    0.567983  0.174267     0.089853   0.443758    0.667935

Apollo.putobs — Function

putobs(x::AbstractArray)

Add an N+1 obervation dimension of size 1 to the tensor x.

Apollo.rgb — Function

rgb(x, lb, ub; bands=[1,2,3])

Visualze the specified bands of x as an rgb image.

Parameters

x: The AbstractArray to be visualized. Must contain at least 3 bands.

lb: A vector of lower bounds for each channel in x. Used by linear_stretch.

ub: A vector of upper bounds for each channel in x. Used by linear_stretch.

bands: A vector of band indices to assign to red, green, and blue, respectively.

Apollo.rmobs — Function

rmobs(x::AbstractArray)

Remove the observation dimension from the tensor x.

Apollo.stackobs — Function

stackobs(x...)
stackobs(x::AbstractVector)

Stack the elements in x as if they were observations in a batch. If x is an AbstractArray, elements will be concatenated along the Nth dimension. Other data types will simply be placed in a Vector in the same order as they were received. Special attention is paid to a Vector of Tuples, where each tuple represents a single observation, such as a feature/label pair. In this case, the tuples will first be unzipped, before their constituent elements are then stacked as usual.

Example

julia> stackobs(1, 2, 3, 4, 5)
5-element Vector{Int64}:
 1
 2
 3
 4
 5

julia> stackobs([(1, :a), (2, :b), (3, :c)])
([1, 2, 3], [:a, :b, :c])

julia> stackobs([rand(256, 256, 3, 1) for _ in 1:10]...) |> size
(256, 256, 3, 10)

julia> xs = [rand(256, 256, 3, 1) for _ in 1:10];

julia> ys = [rand(256, 256, 1, 1) for _ in 1:10];

julia> data = collect(zip(xs, ys));

julia> stackobs(data) .|> size
((256, 256, 3, 10), (256, 256, 1, 10))

Apollo.todevice — Function

todevice(x)

Copy x to the GPU.

Apollo.unzip — Function

unzip(x::AbstractVector{<:Tuple})

The reverse of zip.

Example

julia> zip([1, 2, 3], [:a, :b, :c]) |> collect |> unzip
([1, 2, 3], [:a, :b, :c])

Apollo.vec2array — Function

vec2array(x::AbstractVector, to::AbstractArray, dim::Int)

Reshape the vector x to have the same number of dimensions as to. Missing dimensions are added as singletons while the dimension corresponding to dim will be filled with the values of x.

Apollo.zeros_like — Function

zeros_like(x::AbstractArray)

Construct an array of zeros with the same size and element type as x.

Data Views

Apollo.AbstractView — Type

Super type of all iterators.

Apollo.JoinedView — Type

JoinedView(data...)

An object that iterates over each element in the iterators given by data as if they were concatenated into a single list.

Apollo.MappedView — Type

MappedView(f, data)

An iterator which lazily applies f to each element in data when requested.

Apollo.ObsView — Type

ObsView(data, indices)

Construct an iterator over the elements specified by indices in data.

Apollo.TileSampler — Type

TileSampler(raster, tilesize::Int; stride=tilesize)

An object that iterates over tiles cut from a given raster.

Parameters

raster: An AbstractRaster or AbstractRasterStack to be cut into tiles.

tilesize: The size (width and height) of the tiles.

stride: The horizontal and vertical distance between adjacent tiles.

Apollo.TileSeq — Type

TileSeq(tiles)

An object that iterates over a Vector of tiles.

Apollo.dropobs — Function

dropobs(data, obs::AbstractVector{Int})

Remove all observations from data whose index corresponds to those in obs.

Apollo.filterobs — Function

filterobs(f, data)

Remove all observations from data for which f returns false.

Apollo.mapobs — Function

mapobs(f, data)

Lazily apply the function f to each element in data.

Apollo.repeatobs — Function

repeatobs(data, n::Int)

Create a new view which iterates over every element in data n times.

Apollo.sampleobs — Function

sampleobs([rng], data, n)

Randomly sample n elements from data without replacement. rng may be optionally provided for reproducible results.

Apollo.shuffleobs — Function

shuffleobs([rng], data)

Randomly shuffle the elements of data. Provide rng for reproducible results.

Apollo.splitobs — Function

splitobs(data; at=0.8, shuffle=true)

Return a set of indices that splits the given observations according to the given break points.

Arguments

data: Any type that implements Base.length().

at: The fractions at which to split data.

shuffle: If true, shuffles the indices before splitting.

Example

julia> splitobs(1:100, at=(0.7, 0.2), shuffle=false)
3-element Vector{Vector{Int64}}:
 [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  61, 62, 63, 64, 65, 66, 67, 68, 69, 70]
 [71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90]
 [91, 92, 93, 94, 95, 96, 97, 98, 99, 100]

Apollo.takeobs — Function

takeobs(data, obs::AbstractVector{Int})

Take all observations from data whose index corresponds to obs while removing everything else.

Apollo.zipobs — Function

zipobs(data...)

Create a new iterator where the elements of each iterator in data are returned as a tuple.

Example

julia> zipobs(1:5, 41:45, rand([:cat, :dog], 5)) |> collect
5-element Vector{Any}:
 (1, 41, :dog)
 (2, 42, :cat)
 (3, 43, :cat)
 (4, 44, :dog)
 (5, 45, :cat)

Transforms

Apollo.DType — Type

The super-type of all data types.

Apollo.Image — Type

Image([name])

Represents a type consisting of image data.

Apollo.Mask — Type

Mask([name])

Represents an instance of mask data.

Apollo.AbstractTransform — Type

Super type of all transforms.

Apollo.RandomTransform — Type

Super type of all random transforms.

Apollo.ComposedTransform — Type

ComposedTransform(transforms...)

Apply transforms to the input in the same order as they are given.

Example

julia> r = Raster(rand(256,256, 3), (X,Y,Band));

julia> t = Resample(2.0) |> Tensor();

julia> apply(t, Image(), r, 123) |> size
(512, 512, 3, 1)

julia> apply(t, Image(), r, 123) |> typeof
Array{Float32, 4}

Apollo.Crop — Type

Crop(size::Int)
Crop(size::Tuple{Int,Int})

Crop a tile equal to size out of the input array with an upper-left corner at (1,1).

Apollo.DeNormalize — Type

DeNormalize(μ, σ; dim=3)

Denormalize the input array with respect to the specified dimension. Reverses the effect of normalize.

Parameters

μ: A Vector of means for each index in dim.

σ: A Vector of standard deviations for each index in dim.

dim: The dimension along which to denormalize the input array.

Apollo.FilteredTransform — Type

FilteredTransform(dtype::DType, t::AbstractTransform)

Modify the transform t so that it will only be applied to inputs whose type and name matches dtype. The * operator is overloaded for convenience.

Example

julia> r = Raster(rand(256,256, 3), (X,Y,Band));

julia> t = Image(:x2) * Resample(2.0);

julia> apply(t, Image(), r, 123) |> size
(256, 256, 3)

julia> apply(t, Image(:x2), r, 123) |> size
(512, 512, 3)

julia> apply(t, Mask(:x2), r, 123) |> size
(256, 256, 3)

Apollo.Normalize — Type

Normalize(μ, σ; dim=3)

Normalize the input array with respect to the specified dimension so that the mean is 0 and the standard deviation is 1.

Parameters

μ: A Vector of means for each index in dim.

σ: A Vector of standard deviations for each index in dim.

dim: The dimension along which to normalize the input array.

Apollo.RandomCrop — Type

RandomCrop(size::Int)
RandomCrop(size::Tuple{Int,Int})

Crop a randomly placed tile equal to size from the input array.

Apollo.Resample — Type

Resample(scale)

Resample x according to the specified scale. Mask types will always be resampled with :near interpolation, whereas Images will be resampled with either :bilinear (scale > 1) or :average (scale < 1).

Parameters

x: The raster/stack to be resampled.

scale: The size of the output with respect to the input.

Apollo.Tensor — Type

Tensor(;precision=Float32, layerdim=Band)

Convert raster/stack into a tensor with the specified precision. See tensor for more details.

Parameters

precision: Any AbstractFloat to use as the tensor's type (default = Float32).

layerdim: RasterStacks will have their layers concatenated along this dimension.

Apollo.apply — Function

apply(t::AbstractTransform, dtype::DType, data, seed)

Apply the transformation t to the data of type dtype with the random seed seed.

Apollo.crop — Function

crop(x::AbstractArray, size, ul=(1,1))

Crop a tile equal to size out of x with an upper-left corner defined by ul.

Apollo.denormalize — Function

denormalize(x::AbstractArray, μ::AbstractVector, σ::AbstractVector; dim=3)

Denormalize the input array with respect to the specified dimension. Reverses the effect of normalize.

Parameters

μ: A Vector of means for each index in dim.

σ: A Vector of standard deviations for each index in dim.

dim: The dimension along which to denormalize the input array.

Apollo.normalize — Function

normalize(x::AbstractArray, μ::AbstractVector, σ::AbstractVector; dim=3)

Normalize the input array with respect to the specified dimension so that the mean is 0 and the standard deviation is 1.

Parameters

μ: A Vector of means for each index in dim.

σ: A Vector of standard deviations for each index in dim.

dim: The dimension along which to normalize the input array.

Apollo.raster — Function

raster(tensor::AbstractArray, dims::Tuple; missingval=0)

Restore the raster dimensions given by dims to the provided tensor. The final dimension of tensor, which is assumed to be the observation dimension, will be dropped.

Apollo.resample — Function

resample(x::AbstractRaster, scale::AbstractFloat, method=:bilinear)
resample(x::AbstractRasterStack, scale::AbstractFloat, method=:bilinear)

Resample x according to the given scale and method.

Parameters

x: The raster/stack to be resampled.

scale: The size of the output with respect to the input.

method: One of :near, :bilinear, :cubic, :cubicspline, :lanczos, or :average.

Apollo.resize — Function

resize(x::AbstractRaster, newsize, method=:bilinear)
resize(x::AbstractRasterStack, newsize, method=:bilinear)

Resize the raster/stack x to newsize under the specified method.

Parameters

x: The array to be resized.

newsize: The width and height of the output as a tuple.

method: One of :near, :bilinear, :cubic, :cubicspline, :lanczos, or :average.

Apollo.tensor — Function

tensor([precision], xs...; kwargs...)
tensor([precision], x::AbstractArray; kwargs...)
tensor([precision], x::AbstractRasterStack; layerdim=Band)

Convert one or more arrays to a tensor with an element type of precision. AbstractRasters will be reshaped as necessary to enforce a dimension order of (X,Y,Z,Band,Ti) before adding an observation dimension. Multiple arrays will be concatenated along the observation dimension after being converted to tensors.

Parameters

precision: Any AbstractFloat to use as the tensor's type (default = Float32).

x: One or more AbstractArrays to be turned into tensors.

layerdim: AbstractRasterStacks will have their layers concatenated along this dimension

before being turned into tensors.

Apollo.transform — Function

transform(t::AbstractTransform, dtype::DType, x)
transform(t::AbstractTransform, dtypes::Tuple, x::Tuple)

Apply the transformation t to the input x with data type dtype.

Apollo.upsample — Function

upsample(x::AbstractArray, scale, method=:bilinear)

Upsample the array x according to the given scale and method.

Parameters

x: The array to be upsampled.

scale: The size of the output with respect to the input.

method: One of :linear, :bilinear, :trilinear, or :nearest.

Metrics

Apollo.AbstractMetric — Type

Abstract supertype of all metrics.

Metrics are measures of a model's performance, such as loss, accuracy, or squared error.

Each metric must implement the following interface:

name(::Type{metric}): Returns the human readable name of the metric.

init(metric): Returns the initial state of the metric as a NamedTuple.

update(metric, state, ŷ, y): Returns the new state given the previous state and a batch.

compute(metric, state): Computes the metric's value for the current state.

Example Implementation

struct Accuracy <: ClassificationMetric end

name(::Type{Accuracy}) = "accuracy"

init(::Accuracy) = (correct=0, total=0)

function update(::Accuracy, state, ŷ, y)
    return (correct = state.correct + sum(ŷ .== y), total = state.total + length(ŷ))
end

compute(::Accuracy, state) = state.correct / max(state.total, 1)

Apollo.ClassificationMetric — Type

Super type of all classification metrics.

Apollo.RegressionMetric — Type

Super type of all regression metrics.

Apollo.Metric — Type

Metric(measure::AbstractMetric)

Metric objects are used to store the state for a given AbstractMetric.

Apollo.Accuracy — Type

Accuracy()

Measures the model's overall accuracy as correct / total.

Apollo.MIoU — Type

MIoU(classes::Vector{Int})

Mean Intersection over Union (MIoU) is a measure of the overlap between a prediction and a label. This measure is frequently used for segmentation models.

Apollo.MSE — Type

MSE()

Tracks Mean Squared Error (MSE) as ((ŷ - y) .^ 2) / length(y).

Apollo.Loss — Type

Loss(loss::Function)

Tracks the average model loss as total_loss / steps

Apollo.compute — Function

compute(m::AbstractMetric, state)

Compute the performance measure from the current state.

Apollo.evaluate — Function

evaluate(model, data, measures...)

Evaluate the model's performance on the provided data.

Parameters

model: A callable that takes a single batch from data and returns a tuple of the form (ŷ, y).

data: An iterable of (x, y) values.

measures: A set of AbstractMetrics to use for evaluating model.

Returns

A NamedTuple containing the performance metrics for the given model.

Example

evaluate(DataLoader((xsampler, ysampler)), Accuracy(), MIoU(2)) do (x, y)
    ŷ = model(x) |> Flux.sigmoid
    return (ŷ, y)
end

Apollo.init — Function

init(m::AbstractMetric)

Returns the initial state of the performance measure, which will be subsequently updated for each mini-batch of labels and predictions.

Apollo.name — Function

name(m::AbstractMetric)
name(m::Metric)

Human readable name of the given performance measure.

Apollo.reset! — Function

reset!(metric::Metric)

Reset the metric's state.

Apollo.update — Function

update(m::AbstractMetric, state, ŷ, y)

Return the new state for the given batch of labels and predictions.

Apollo.update! — Function

update!(metric::Metric, ŷ, y)

Update the metric state for the next batch of labels and predictions.

Performance Tracking

Apollo.Tracker — Type

Tracker(metrics...)

An object to track one or more metric values over the course of a training run. Each metric can be either an AbstractMetrics or a name => metric pair. In the first case, the default name of the provided metric will be used, while the second allows us to choose an arbitrary name.

Tracker implements the Tables.jl interface, allowing it to be used as a table source.

Example 1

julia> tracker = Tracker(Accuracy(), MIoU([0,1]));

julia> step!(tracker, [0, 0, 1, 0], [0, 1, 1, 0]);  # update all metrics

julia> scores(tracker)
(epoch = 1, accuracy = 0.75, MIoU = 0.5833333730697632)

julia> epoch!(tracker);

julia> tracker
Tracker(current_epoch=2)
┌───────┬──────────┬──────────┐
│ epoch │ accuracy │     MIoU │
│ Int64 │  Float64 │  Float64 │
├───────┼──────────┼──────────┤
│     1 │     0.75 │ 0.583333 │
│     2 │      0.0 │      0.0 │
└───────┴──────────┴──────────┘

Example 2

julia> tracker = Tracker("train_acc" => Accuracy(), "val_acc" => Accuracy());

julia> step!(tracker, "train_acc", [0, 0, 1, 0], [0, 1, 1, 0]);  # specify metric to update

julia> step!(tracker, r"val_", [0, 0, 0, 0], [0, 1, 1, 0]);  # update metrics matching regex

julia> tracker
Tracker(current_epoch=1)
┌───────┬───────────┬─────────┐
│ epoch │ train_acc │ val_acc │
│ Int64 │   Float64 │ Float64 │
├───────┼───────────┼─────────┤
│     1 │      0.75 │     0.5 │
└───────┴───────────┴─────────┘

Apollo.MetricDict — Type

MetricDict(metrics...; prefix="")

An object to track one or more metrics. Each metric is associated with a unique name, which defaults to name(metric). This can be overriden by providing a name => metric pair. The prefix keyword adds a constant prefix string to every name.

Example 1

julia> md = MetricDict(Accuracy(), MIoU([0,1]); prefix="train_");

julia> step!(md, [0, 0, 1, 0], [0, 0, 1, 1]);

julia> md
MetricDict(train_accuracy=0.75, train_MIoU=0.5833334)

julia> step!(md, [0, 0, 1, 1], [0, 0, 1, 1]);

julia> md
MetricDict(train_accuracy=0.875, train_MIoU=0.775)

julia> reset!(md)

julia> md
MetricDict(train_accuracy=0.0, train_MIoU=0.0)

Example 2

julia> md = MetricDict("train_acc" => Accuracy(), "val_acc" => Accuracy());

julia> step!(md, "train_acc", [0, 1, 1, 0], [1, 1, 1, 0]);  # update train_acc

julia> step!(md, r"val_", [1, 1, 1, 0], [1, 1, 1, 0]);  # update metrics matching regex

julia> md
MetricDict(train_acc=0.75, val_acc=1.0)

Apollo.Max — Type

Max(metric::String)

Represents an ordering of largest to smallest for the given metric.

Apollo.Min — Type

Min(metric::String)

Represents an ordering of smallest to largest for the given metric.

Apollo.best_epoch — Function

best_epoch(tracker::Tracker, order::Order)

Returns the best epoch in tracker according to the given order.

Apollo.current_epoch — Function

current_epoch(tracker::Tracker)

The current epoch in tracker.

Apollo.epoch! — Function

epoch!(tracker::Tracker)

Store the metric value for the current epoch and reset the state.

Apollo.printscores — Function

printscores(t::Tracker; epoch=current_epoch(t), metrics=r".*")

Return the metric scores for the provided epoch as a pretty-printed string.

Apollo.scores — Function

scores(t::Tracker; epoch=current_epoch(t), metrics=keys(t.metrics))

Return the metric scores for the provided epoch.

Apollo.step! — Function

step!(x::MetricDict, ŷ, y)
step!(x::MetricDict, metric::String, ŷ, y)
step!(x::MetricDict, metric::Regex, ŷ, y)

Update the metric for the current epoch using the provided prediction/label pair.

step!(tracker::Tracker, ŷ, y)
step!(tracker::Tracker, metric::String, ŷ, y)
step!(tracker::Tracker, metric::Regex, ŷ, y)

Update the metric for the current epoch using the provided prediction/label pair.

Layers

Apollo.ConvBlock — Function

ConvBlock(filter, in, out, σ; depth=2, batch_norm=true)

A block of convolutional layers with optional batch normalization.

Parameters

filter: The size of the filter to use for each layer.

in: The number of input features.

out: The number of output features.

σ: The activation function to apply after each convolution.

depth: The number of successive convolutions in the block.

batch_norm: Applies batch normalization after each convolution if true.

Apollo.LSTM — Type

LSTM(in => out)

A stateless pixel-wise LSTM layer that expects an input tensor with shape WHCLN, where W is image width, H is image height, C is image channels, L is sequence length, and N is batch size. The LSTM module will be applied to each pixel in the sequence from first to last with all but the final output discarded.

Parameters

in: The number of input features.

out: The number of output features.

Example

julia> m = Apollo.LSTM(2=>32);

julia> x = rand(Float32, 64, 64, 2, 9, 4);

julia> m(x) |> size
(64, 64, 32, 4)

Apollo.SeparableConv — Function

SeparableConv(filter, in, out, σ=Flux.relu)

A separable convolutional layer consists of a depthwise convolution followed by a 1x1 convolution.

Parameters

filter: The size of the filter as a tuple.

in: The number of input features.

out: The number of output features.

σ: The activation function to apply following the 1x1 convolution.

Encoders

Apollo.AbstractEncoder — Type

AbstractEncoder{F1,F2,F3,F4,F5}

Super type of all encoder models. The type parameters F1, F2, F3, F4, and F5 denote the number of features output by each block of the encoder.

Apollo.ResNet18 — Type

ResNet18(;weights=nothing)

Construct a ResNet18 encoder with the specified initial weights.

Parameters

weights: Either :ImageNet or nothing.

Apollo.ResNet34 — Type

ResNet34(;weights=nothing)

Construct a ResNet34 encoder with the specified initial weights.

Parameters

weights: Either :ImageNet or nothing.

Apollo.ResNet50 — Type

ResNet50(;weights=nothing)

Construct a ResNet50 encoder with the specified initial weights.

Parameters

weights: Either :ImageNet or nothing.

Apollo.ResNet101 — Type

ResNet101(;weights=nothing)

Construct a ResNet101 encoder with the specified initial weights.

Parameters

weights: Either :ImageNet or nothing.

Apollo.ResNet152 — Type

ResNet152(;weights=nothing)

Construct a ResNet152 encoder with the specified initial weights.

Parameters

weights: Either :ImageNet or nothing.

Apollo.StandardEncoder — Type

StandardEncoder()

Construct a standard UNet encoder with no pre-trained weights.

Models

Apollo.Classifier — Type

Classifier(encoder::AbstractEncoder; channels=3, nclasses=1, batch_norm=true)

Construct an image classifier from the provided encoder.

Apollo.DeeplabV3 — Type

DeeplabV3(encoder::AbstractEncoder; channels=3, nclasses=1)

Construct a DeeplabV3+ model.

Apollo.R2UNet — Type

R2UNet(in_features, n_classes; t=2, batch_norm=false)

Construct an R2-UNet model.

Apollo.SSC_CNN — Type

SSC_CNN(;width=128)

Construct an SSC_CNN model (Nguyen et al.) to sharpen the 20m Sentinel-2 bands to a resolution of 10m.

Parameters

width: The number of features to use in each block (default = 128).

Apollo.UNet — Type

UNet(;encoder=StandardEncoder(batch_norm=true), input=nothing, channels=3, nclasses=1, batch_norm=true)

Construct a UNet model.

Keywords

encoder: The encoder to use for the UNet model. Defaults to the standard encoder.

input: The input block, which defaults to two convolutional layers as with standard UNet.

channels: The number of input channels to use when input=nothing. Ignored when input is specified.

nclasses: The number of output channels produced by the head.

batch_norm: Use batch normalization after each convolutional layer (default=true).