Apollo

Documentation for Apollo.

Apollo.AbstractMetric

Apollo.Metric

Apollo.ObsView

Apollo.compute

Apollo.crop

Apollo.evaluate

Apollo.folddims

Apollo.init

Apollo.name

Apollo.splitobs

Apollo.update

Apollo.update!

Apollo.AbstractMetric — Type

Abstract supertype of all evaluation metrics.

Apollo.Metric — Type

Metric(measure::AbstractMetric)

Construct a Metric object to track the state for the given AbstractMetric.

Apollo.ObsView — Type

ObsView(data, indices)

Construct an iterator over data

Apollo.compute — Function

compute(m::AbstractMetric, state)
compute(m::Metric, state)

Compute the performance measure from the current state.

Apollo.crop — Function

crop(x::AbstractArray, size, ul=(1,1))

Crop a tile equal to size out of x with an upper-left corner defined by ul.

Apollo.evaluate — Method

evaluate(model, data, measures...)

Evaluate the model's performance on the provided data.

Parameters

model: A callable that takes a single batch from data and returns a tuple of the form (ŷ, y).

data: An iterable of (x, y) values.

measures: A set of AbstractMetrics to use for evaluating model.

Returns

A NamedTuple containing the performance metrics for the given model.

Example

evaluate(DataLoader((xsampler, ysampler)), Accuracy(), MIoU(2)) do (x, y)
    ŷ = model(x) |> Flux.sigmoid
    return (ŷ, y)
end

Apollo.folddims — Method

folddims(f, xs::AbstractRaster; dim=Band)

Reduce the collection of non-missing values to a singular value in each slice of x WRT dims.

Arguments

f: A function that reduces an array of values to a singular value.

x: An AbstractRaster over which we want to fold.

dims: The dimension used to generate each slice that is passed to f.

Example

julia> μ = folddims(mean, raster, dims=Band)
6-element Vector{Float32}:
 0.09044644
 0.23737456
 0.30892986
 0.33931717
 0.16186203
 0.076255515

Apollo.init — Function

init(m::AbstractMetric)

Returns the initial state of the performance measure, which will be subsequently updated for each mini-batch of labels and predictions.

Apollo.name — Function

name(m::AbstractMetric)
name(m::Metric)

Human readable name of the given performance measure.

Apollo.splitobs — Method

splitobs(data; at=0.8, shuffle=true)

Return a set of indices that splits the given observations according to the given break points.

Arguments

data: Any type that implements Base.length().

at: The fractions at which to split data.

shuffle: If true, shuffles the indices before splitting.

Example

julia> splitobs(1:100, at=(0.7, 0.2), shuffle=false)
3-element Vector{Vector{Int64}}:
 [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  61, 62, 63, 64, 65, 66, 67, 68, 69, 70]
 [71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90]
 [91, 92, 93, 94, 95, 96, 97, 98, 99, 100]

Apollo.update — Function

update(m::AbstractMetric, state, ŷ, y)

Return the new state for the given batch of labels and predictions.

Apollo.update! — Method

update!(metric::Metric, ŷ, y)

Update the metric state for the next batch of labels and predictions.